back/requirements.txt
asyncio
aiohttp
asynctnt
envparse
bcrypt
PyJWT
cchardet
datetime
json

